Account for orientation of GD
Master branch does not account for the orientation of GD when filtering out invariant potential pseudo-QTNs. The current code works when GD is in row orientation but not column orientation. When `length(Porder)` < `nrow(GDneo)` this adds a large number of NA's to the end of `SNP.index` and then to the end of `Porder`. This leads to `seq(bottom:up)` producing a sequence longer than the `ncol(GDneo)`. The proposed change accounts for different orientations of GD in the filtering of invariant pseudo-QTNs.